UPBGE: Fix compilation

+++ source/blender/blenkernel/intern/mball_tessellate.cc
@@ -1186,7 +1186,7 @@ static void init_meta(Depsgraph *depsgraph, PROCESS *process, Scene *scene, Obje
   char obname[MAX_ID_NAME];
   SceneBaseIter iter;
   const eEvaluationMode deg_eval_mode = DEG_get_mode(depsgraph);
-  const short parenting_dupli_transflag = (OB_DUPLIFACES | OB_DUPLIVERTS);
+  const int parenting_dupli_transflag = (OB_DUPLIFACES | OB_DUPLIVERTS);
 
   /* Copy object matrices to cope with duplicators from #BKE_scene_base_iter_next. */
   float obinv[4][4], obmat[4][4];
